Added sound effects, and modularized js functions into a debugginGameUtils subdirectory.
