--- src/js/constants.js
@@ -1,5 +1,0 @@
-export const scaleFactor = 3;
-export const interactionRange = 140;
-export const debug = false;
-
-if (debug) console.log("constants.js is loaded!")
