+++ src/js/debuggingGameUtils/constants.js
@@ -1,0 +1,6 @@
+export const scaleFactor = 3;
+export const interactionRange = 140;
+export const debug = true;
+
+window.debug = debug;
+
