+++ src/js/debuggingGameUtils/index.js
@@ -1,0 +1,9 @@
+export { playerInteractingWith } from "./playerInteraction";
+export { interactionRange, scaleFactor } from "./constants";
+export { handleLayers } from "./handlingLayers";
+export { k } from "./kaboomCtx";
+export { settingUpCameraMovement } from "./settingUpCamera";
+export { mouseMovement } from "./movement";
+export { createPlayer } from "./createPlayer";
+export { displayDialogue } from "./displayDialogue";
+export { loadAssets } from "./loadAssets";
