+++ src/js/debuggingGameUtils/kaboomCtx.js
@@ -1,0 +1,11 @@
+import kaboom from "kaboom";
+
+export const k = kaboom({
+  global: false,
+  touchToMouse: true,
+  canvas: document.getElementById("game"),
+  debug: false, // set to false once ready for production
+  background: [80, 2, 80],
+});
+
+
